python-app 3/requirements.txt
flask==3.0.3
python-dotenv==1.0.1
openai==1.45.0
gunicorn==22.0.0